% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_EP_output_folder.R
\name{clean_EP_output_folder}
\alias{clean_EP_output_folder}
\title{Remove Extra EnergyPlus Output Files}
\usage{
clean_EP_output_folder(directory)
}
\description{
EnergyPlus runs produce many files. This script cleans up the unnecessary ones.
}
\examples{
clean_EP_output_folder("./results/FCZ9_Office-Like_ALL")
}
\keyword{Calibration}
\keyword{EnergyPlus,}
\keyword{Parametrics,}
