% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{clean_RunDirMulti_folder}
\alias{clean_RunDirMulti_folder}
\title{Delete some of the extra files from the folder with the batch executable RunDirMulti.bat}
\usage{
clean_RunDirMulti_folder(path)
}
\arguments{
\item{path}{Folder location relative to working directory}
}
\description{
This function searches for fields in a dataframe representing an EnergyPlus output for units if [J] and then
converts them into units of [kW].
}
\examples{
clean_RunDirMulti_folder(sim_path)
}
\keyword{Calibration}
\keyword{Data,}
\keyword{EnergyPlus,Model}
