% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sections.R
\name{combine_sections}
\alias{combine_sections}
\title{Combine EnergyPlus Input File Sections}
\usage{
combine_sections(building_subcategory, file_00)
}
\arguments{
\item{building_subcategory}{Name of the EnergyPlus run}

\item{file_00}{The first input file section}
}
\description{
At ADM we divide the EnergyPlus input files into various sections to make things easier.
}
\examples{
combine_sections(building_subcategory, file_00)
}
\keyword{Calibration}
\keyword{EnergyPlus,}
\keyword{Parametrics,}
