% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parametric.R
\name{create_parametric}
\alias{create_parametric}
\title{Create Parametric File}
\usage{
create_parametric()
}
\description{
Create Parametric Section of EnergyPlus Input File Section
}
\examples{
combine_sections()
}
\keyword{Calibration}
\keyword{EnergyPlus,}
\keyword{Parametrics,}
