% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_meter_data_by_zone.R
\name{split_meter_by_zone}
\alias{split_meter_by_zone}
\title{Split Meter Data By Zone}
\usage{
split_meter_by_zone(meter_path)
}
\arguments{
\item{meter_path}{Path to meter data}
}
\description{
The meter files by utility are too big and calibrations tend to be done by zone
}
\examples{
split_meter_by_zone()
}
\keyword{Calibration}
\keyword{EnergyPlus,}
\keyword{Parametrics,}
