% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_enduse_coefficients.R
\name{update_enduse_coefficients}
\alias{update_enduse_coefficients}
\title{Update Enduse Coefficients}
\usage{
update_enduse_coefficients(coefficients_path, coefficients_file, file_00)
}
\arguments{
\item{coefficients_path}{Path to coefficients}

\item{coefficients_file}{Path to coefficients}

\item{file_00}{Path to coefficients}
}
\description{
End use coefficients are calculated with a regression approach in the analytical calculator. This function reads them in, 
and updates the EnergyPlus input file with the new values.
}
\examples{
update_enduse_coefficients()
}
\keyword{Calibration}
\keyword{EnergyPlus,}
\keyword{Parametrics,}
