% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_schedule_section_references.R
\name{update_schedule_section_references}
\alias{update_schedule_section_references}
\title{Update Schedule Section References}
\usage{
update_schedule_section_references(building_subcategory)
}
\arguments{
\item{building_subcategory}{Building type}
}
\description{
Note that the parametric file is set up to run multiple schedules.
For now, we are not going to go there. In "02_Schedules.idf" there are
calls to the parametric file using the code "$8760File". Rather than a
parametric run, this script replaces the file with the appropriate analytic
schedule name.
}
\examples{
update_schedule_section_references(schedules_folder, sections_folder, building_subcategory)
}
\keyword{Calibration}
\keyword{EnergyPlus,}
\keyword{Parametrics,}
