% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_setpoint_schedules.R
\name{change_setpoint_schedules}
\alias{change_setpoint_schedules}
\title{Change Setpoint Schedules}
\usage{
change_setpoint_schedules(building_subcategory, research_schedule_name)
}
\arguments{
\item{research_schedule_name}{A schedule of setpoints to try}

\item{building_subctegory}{Building Subcategory}
}
\description{
Replaces schedule columns with calculated values
}
\examples{
change_setpoint_schedules(building_subcategory, setpoint_research.csv)
}
\keyword{Calibration}
\keyword{EnergyPlus,}
\keyword{Parametrics,}
